+++ docs/source/release_notes.rst
@@ -5,7 +5,7 @@ Release 0.2.0 (2024-04-15)
 --------------------------
 
 - GEOPY-1420: Add Documentation.
-- GEOPY-1432: Drillhole name first in exported las files.
+- GEOPY-1432: Drillhole name first in exported LAS files.
 - GEOPY-1414: Suffix exported .las files with property group name.
 - GEOPY-1413: Move all parameters in import ui.json to main tab.
 - GEOPY-1320: Expose collocation tolerance in import ui.json file